[ADD] frontdesk: visitor management app
PURPOSE
========

Visitor management helps an organization keep track of the people who
visit their location. A visitor management system makes the check-in process
painless. When guests sign in, your digital visitor log updates in real-time.
Monitor it daily, or know it's there when you need it. For instance, once the
visitor enters the organization, the digital check-in system boosts the
visitor's confidence and well-being by providing a guide and ready directions,
building on the professional first impression.

SPECIFICATION
=============

Backend
-----------

Station Configuration:

In the backend, the "Frondesk" module offers comprehensive station management
capabilities. When creating a station, administrator can input essential
details such as the station name, set the responsibles for the station, and add
the description to show on the frontend. Each station is provided with a
dedicated URL for frontend access, including a unique secret key. Additionally,
station image and the ability to set the theme for frontend representation are
integral features.

Also, administrator and station responsible can configure various options like:
host selection - which allows the visitor to pick the host.
drink options - allows the visitor to select a drink during registration.
self-check-in - which shows a QR code on the screen to check in from mobile.
authentify guest - request additional information upon registering.
notifications - The host and station responsible will receive a notification
when a visitor registers.

Station view is facilitated through both Kanban and list views, offering
efficient management options. The kanban view provides quick insights into
displaying the number of planned visitors, visitors on site, drinks to serve,
along with the time of the latest check-in visitor.

Visitor:

The backend functionalities of the module extend to visitor management, ensuring
seamless creation, updates, and removal of visitors. Various views, including
Gantt view, kanban view, calendar view, pivot view, graph view, and list view
are available for efficient visitor management.

Additionally, administrator and responsible of the station can create planned
visitors from the backend, which helps the visitors to quick check in from the
frontend.

Frontend
-------------

Visitor Check-In:

Desk managers can efficiently check in visitors through the frontend, ensuring a
seamless process. During the registration process, the frontend offers a
selection of drinks for visitors to choose from, enhancing their overall
experience.

QR Code Check-In:

Visitors can speed up the check-in process by using QR codes displayed on the
Welcome Screen, significantly reducing wait times.

Quick Sign-In:

If a visitor has been pre-registered from the backend with a scheduled check-in
time. This innovative feature ensures efficiency by presenting planned visitors
approximately 45 minutes before and after their designated check-in time.
Whether visitors arrive early or late, the "Quick Sign-In" panel offers a
seamless check-in experience, minimizing wait times and enhancing overall
visitor satisfaction.

TECHNICAL SPECIFICATION
=======================

Models
----------

frontdesk_frontdesk model:

The model is dedicated to managing stations. It likely contains fields such as
station names, station responsibles, messages to show on the frontend, options
like host selection - which allows the visitor to pick the host, drink options -
allows the visitor to select a drink during registration, self-check-in - which
shows a QR code on the screen to check in from mobile, authentify guest -
request additional information upon registering, notifications - The host and
station responsible will receive a notification when a visitor registers, theme
and image information for the frontdesk website.

frontdesk_visitor model:

The frontdesk_visitor model is designed for the management of visitors. It
likely contains fields such as visitor name, phone number, email, company,
check-in and check-out time, duration, station, drinks, host, message for the
visitor and visitor's status.

frontdesk_drink model:

This model represents drinks that can be configured from the station. It
contains fields like name, image, sequence, and people to notify - It will
notify the people once a visitor selects a drink.

Access Rights
------------------

In security, we have 2 groups.

Administrator Group:

Members of the "Administrator" group possess comprehensive access rights. They
are granted the ability to: Create, update, and delete stations, visitors as
well as drinks.

User Group:

The "User" group, on the other hand, has more limited permissions.

Members of this group are configured with the following access rights:
Read and update rights for the station for which they are responsible.
We are allowing user groups to create, read, and update rights for their
visitors.

Task-3381939

closes odoo/enterprise#45725

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Earth Patel <eapa@odoo.com>
Co-authored-by: Shaan Thakkar <thsh@odoo.com>
Co-authored-by: Ipsita Borisagar <ibo@odoo.com>