[IMP] l10n_cl_edi_exports, l10n_mx_edi: incoterm
In those EDI some xpath were broken due to change in the community pr
(https://github.com/odoo/odoo/pull/118954).
For l10n_mx, changing the xpath was enough.
For l10n_cl the xpath is useless since the information of incoterm in repeated
twice (header and bottom) and we decided to remove it from the bottom of the
invoice

closes odoo/enterprise#43363

Task: 3273460
Related: odoo/odoo#118954
Related: odoo/upgrade#4598
Signed-off-by: William André (wan) <wan@odoo.com>