[FIX] industry_fsm: only make basic projects billable
Before this commit, since #36459, the partner field in project and
task is hidden if the project is non-billable or not fsm one.
When Sales app is installed, all projects with a partner set or
with at least one of their tasks with a partner set, the projects
are made billable in `sale_project` module.

This commit filters the projects to only make billable the basic
projects and not the fsm projects.

closes odoo/enterprise#43384

X-original-commit: f7c1e341d82b5af906cab41ad847cb87b1d9c770
Related: odoo/odoo#126880
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>