[FIX] delivery_sendcloud: skip code if 'access_point_address' is missing
Steps to reproduce the bug:
- Install delivery sendcloud.
- Create a new shipping method with "sendcloud" as the provider.
- Create a sale order:
- Add any storable product.
- Click on "add shipping".
- Select "sendcloud".
- Confirm the sale order.
- Go to the linked delivery.
- Try to validate it.

Problem:
A traceback is triggered: "AttributeError: 'sale.order' object has no
attribute 'access_point_address'".

The field "access_point_address" is defined in the "Website_sale"
module, and the "delivery_sendcloud" module does not depend on it:
https://github.com/odoo/enterprise/blob/88d28042fde35811bf01ee012bd245ade6289885/delivery_sendcloud/__manifest__.py#L10

Solution:
Check if the field is present in the record before using it to ensure
that "website_sale" is installed.

opw-3389478

closes odoo/enterprise#43603

X-original-commit: 68a12afd6b39d02840bad64713b4e4c631bc66c1
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>