[IMP] timesheet_grid: improve color code used for represent worked time
before this commit, color codes used for represent overtime, exact time and
missing hour are not consistent with everywhere else.
make changes in color code to get better understanding of worked time. highlight
negative overtime in red, highlight day total in orange if there is overtime,
highlight the total time for the period in red, green and orange according to
overtime.

task-3251683

closes odoo/enterprise#43493

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>