[IMP] spreadsheet: use company currency as default format
Currently, the currency in menu
"Format > Number > Currency" (or rounded currency) is dollar ($).

If you live in Europe and want to use €, you always have to use the custom
currency format option.

This commit allows to configure the default currency proposed in this
menu. The main company currency is used.

Task 3389619

closes odoo/enterprise#43425

Related: odoo/odoo#126933
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>