[FIX] account_reports_cash_basis: fix bad forward-port of test
'cash_basis' option key becomes 'report_cash_basis' in 16.0. This was wrongly forward-ported from the original 15.0 fix in https://github.com/odoo/enterprise/commit/f2833e48feefe7be7d0c9f99de9f447bc3b4af97.

Putting 'cash_basis' in the options had no effect. The test was only considering one invoice, with a payment made prior to it: in this case, the invoice has to appear at the invoice date ... which is exactly what will happen when not using the cash basis feature. So, in this case, the report was called without the cash basis, but returned the expected result. We make this more robust by creating a second invoice without any payment at the same date as the existing one: this invoice will not appear in the report when using the cash basis method, ensuring the result of the report will be different with or without the option.

closes odoo/enterprise#43443

Signed-off-by: William André (wan) <wan@odoo.com>