[FIX] sale_subscription: sale_order_template_id was removed upon save
The sale_order_template_id field is displayed twice in the sle order view.
Before this commit, the second occurence has a readonly attribute. As Odoo uses
the attribute modifiers of the last encoutered field, it would assume that the
field was radonly and prevent keeping the changes made on the field.
The sale_order_template_id would be reset upon save.

Taskid: 3258616
Part-of: odoo/enterprise#39195