[ADD] stock_barcode_quality_mrp: to make quality check effective in mrp barcode
Before this commit
==================
Currently, The quality check button was not visible in the barcode when there is
quality checks to do for manufacturing orders.

After this commit
=================
With this commit enable quality check button and disable validation button on
production with barcode when there still quality checks to do on the manufacturing order.

task - 3358263

closes odoo/enterprise#43447

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>