[FIX] knowledge: rewords PermissionPanel's restricting access dialog
This commit rewords the confirm button for the restricting access dialog of
the PermissionPanel.
The confirm button is now 'Restrict access' when restricting the access of
another user and 'Restrict own access' when the user is restricting its
permission on the article.
We also updated the Knowledge pot file since we added a new translated term.

task-3381759

closes odoo/enterprise#43752

X-original-commit: 15eda3969a6affda29779354f1a9e17110270690
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>