[FIX] planning: hide user_id field of resource form
Before this commit in planning list view when we create & edit a new
resource  on the fly user_id is shown which is not case when we do same thing from
planning gantt view dialog.

After this commit in planning list view and gantt view dialog's resource field will
be ideal.

task - 3217930

closes odoo/enterprise#43534

X-original-commit: 297092d4394d740fc771098971e793181f0193c9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>