[IMP] knowledge: add nocontent helper to misconfigured item calendar
Purpose:
--------
If the property used as start date of an item calendar has been deleted,
or if the user opened the item calendar in full screen and reloaded the
page, the calendar will be empty (because either the records do not have
the deleted property anymore, or because when the page is reloaded, the
itemCalendarProps are lost since they are behaviorProps.

This commit adds a nocontent helper message on top of the item calendar
in  these cases, so that the user knows that it is empty because it is
misconfigured.

Task-3239985

Part-of: odoo/enterprise#40993