[FIX] project_forecast: display `Planned` stat button in project update
Before this commit since #40050, the `Planned` was no longer visible
in project update because `show` property is not given in the object
and so the system the button has invisible.

This commit adds the property to be able to show that stat button
as before in the right side panel of project update view.

closes odoo/enterprise#43520

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>