[FIX] documents: prevent crash caused by translating technical values
The `t-set` directive can be used with a body value, in which case it is
supposed to be translated. This was not the case until a recent owl fix
(2.1.4), but now that Owl is fixed, the translation is properly applied
and the document inspector code now crashes in non-english languages,
since the code is using a field name as a key in an object.

To reproduce the crash: install another language, french for example,
activate it, go to documents and open a document => it crashes before
this commit.

Note that I just fixed quickly all obvious occurences, but in general,
no technical data should be translated.

opw-3394696

closes odoo/enterprise#43522

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>