[FIX] l10n_cl_edi: extra fromstring tostring needed anyways
Because of PR https://github.com/odoo/enterprise/pull/37891
while changing the Markup in the method for signing the invoices,
we also removed one etree fromstring/tostring and the functional
tests for the signing seemed to pass with it.

However, this extra fromstring/tostring made the signing invalid
and made it impossible to sign invoices anymore.

By putting back the extra fromstring/tostring to signing passes
again.  The difference between both fromstring/tostring is that the
second does the c14n.

opw-3395688
opw-3395815
opw-3396472
opw-3396644
opw-3396622
opw-3396893
opw-3396898
opw-3396915

Part-of: odoo/enterprise#43564