[FIX] mrp_workorder: component history loss in tablet view
Before this commit:
====================
In tablet view when a register component check has no
stock.move.line(e.g. not reserved). During the record production the quantity
on the check will be set to 0. And the registered quantity will be loss on the
check (but exists on sml).

After this commit:
====================
We never write the quantity on the check during the extra move line process.

task-3212121

closes odoo/enterprise#44077

X-original-commit: 9bb33a987e7cbefba6e320b78d29f34f76ed379f
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>