[FIX] mrp_workorder: component history loss in tablet view
Before this commit:
====================
In tablet view when a register component check has no
stock.move.line(e.g. not reserved). During the record production the quantity
on the check will be set to 0. And the registered quantity will be loss on the
check (but exists on sml).

After this commit:
====================
We never write the quantity on the check during the extra move line process.

task-3212121

closes odoo/enterprise#43518

X-original-commit: 9e645932deef03b8b599d6ed5b231f566371bb03
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>