[FIX] web_studio: fix color of toggle Studio button
This commit changes the toggler button to an actual button element, which
allows the icon to be displayed as expected in black or white, depending
the actual theme used for Odoo. The alt attribute has been removed, since
it is not supported on a <button> element, neither it was on the previously
used <a> element.

By using an <a> element, the font was colored in cyan instead. Other systray
items were already using <button> elements.

Tests and selectors have also been adapted to this change.

closes odoo/enterprise#43624

X-original-commit: d9beb80e0dcc7e1ffe2554c3e4f547e6cc7aac6b
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>