[REV] account_asset: don't create moves before lock date
In commit 55f3bb73ea0360bc3b2bae246ce0a6918f5f5512, when recomputing
the draft moves of an asset depreciation, we check the lock date and
adapt the dates of these draft moves to be after the lock date.

However we could have set a wrong lock date and later change the
lock date to an earlier date. The draft depreciation moves might now be
posted on a later date then we want according to the new lock date.

opw-3305506

Part-of: odoo/enterprise#43586