[FIX] account_asset: test unlinking draft moves before lock date
Currently when unlinking draft depreciation moves, we check whether the
dates of these moves are not before a lock date. If so, we show an error
to the user.

However, if the move is not posted, we have no reason to show an error,
since unlinking an unposted move, even before a lock date, is a valid
operation. The related change in community fixes this issue. Here we
test whether we can successfully change depreciation methods and thus
unlink draft moves before a lock date.

opw-3305506

closes odoo/enterprise#44199

Related: odoo/odoo#128701
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>