[FIX] hr_contract_salary, *: fix demo data
*: sale_renting_sign

Steps to reproduce:

- login as admin
- install sale_renting_sign
- activate debug mode
- sign > 3 dots menu on Rental Agreement > properties
	- Who can cign: All Users
	- Template access group: Sign / User: Own and Shared Templates
	- Save

- logout and login with user who has Sign / User privilege
- Rental Agreement > Send

Issue:

User gets error message: `Sorry, you are not allowed to access this
document.` If we repeat the same process on another existing template,
such as `Non-Disclosure Agreement.pdf`, the user will not get an error
when clicking Send.

This happens because the `check` function will raise an error if the
template does not have a `res_model` and `res_id`.

opw-3250596

closes odoo/enterprise#43899

X-original-commit: e2b77b70c575900ffc2948829cd32808c1b6e8c1
Signed-off-by: Kevin Baptiste <kba@odoo.com>