[IMP] knowledge: move delete button at bottom
When the user wants to perform any action on the article (from the top right
corner, there is an action menu), the delete button is between the other
buttons. So, sometimes users maybe misclick on the delete button.

With this commit, the delete button moved to the bottom of the action buttons.

Task-3397730

closes odoo/enterprise#43655

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>