[FIX] account_asset: account change on depreciation results in 0
To reproduce:
- Create an asset
- Compute the board
- On an entry, change the expense account
- The depreciation is now 0 instead of the value

The depreciation should not be 0 as the asset
is still depreciated.
We now compute the amount based on the amount
of the right type instead of just a single account.

A lot of the tests have changed because they were
putting the wrong account type for the account fields of
the assets, and so, testing things that should not be done
(and blocked from the interface).
Also did some "cosmetic" changes on tests, so we have variables names consistent with the asset's type

Also permit changing the account on the asset even when there is a lock date, to only change the account of the moves that can be changed.
task-3377189

closes odoo/enterprise#45218

X-original-commit: 6673524d4e72e2b1e44de0c324df4d092207d243
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>