[IMP] web_enterprise: add share option from user menu
This commits adds a new share entry from the user menu. This allows the
share of any web page from the PWA (standalone mode -> URL is not visible).

On smaller screens, the share is present directly at the top of the Burger
Menu instead of being an item of the list.

Tests have been added to verify the presence of the button when Odoo is
running in standalone mode, as well as the visibility depending the screen
size.

task-3341052

closes odoo/enterprise#43680

Related: odoo/odoo#128969
Signed-off-by: Pierre Paridans (app) <app@odoo.com>