[FIX] sale_{planning,project_forecast}: add project in domain to fetch shifts
Before this commit, when the user goes to project update of a Project A
and click on `Planned` stat button to display the shifts linked to
that Project A. Then the user will click on `Plan Orders` button,
a traceback will occurs saying `is_closed` field does not exist in
`project.task` model.

This commit fixes the issue by changing `('is_closed', '=', False)`
in the domain used into `('state', 'not in', ['1_done', 1_canceled])`
since `is_closed` has been recently removed. Also, this commit moves
that code in sale_project_forecast because `project.task` model does not
exist in `sale_planning` module since `project` module is not in the
dependencies of that module.

Bug found duing the development of task-3277191

closes odoo/enterprise#43629

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>