[FIX] mrp_accountant: fix dependencies
The module mrp_accountant explicitly inherits methods and a view from
stock_accountant but is not directly dependant on it. The dependency is
implicit by the auto_install of stock_accountant and the shared
dependencies the modules have. However the auto_install can be bypassed
by users post installation. The dependency should be explicit.

closes odoo/enterprise#43939

X-original-commit: 5ecb96a60e904f8d5b68f551cad0b1fb81340e62
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>