[IMP] l10n_mx_edi: improve payment way/method fields
Problem
---------

At the moment, the "Uso" field is present on the Sale Order, but not the
Payment Way. Which means that everytime a sale order is creating an
invoice, the invoice must be edited to change the Payment Way. This is
especially annoying for recurring sale orders like subscriptions.

Additionally, you will usually use the same values by default when
selling or invoicing the same customer. It would then save a lot of time
to be able to register default values for those fields at the Customer
level.

Objective
---------

- 'Payment way' now updatable in sale_order and the set value is used
when creating an invoice.
- 'Payment way' and 'Uso' are now setable on the customer sheet; the
values are used as default when creating a sale order and/or invoice.

Solution
---------

Use compute function to set the values, values are recomputed everytime
the costumer is modified on the sale order or invoice. The compute
functions read the partner_id relevant field values to check if they are
set.

task-3388110

closes odoo/enterprise#43625

Signed-off-by: Laurent Smet (las) <las@odoo.com>