[IMP] timesheet_grid: hide Add line btn for locked periods
Before this commit:
My/All Timesheets Grid View: Add new line button is displayed when the current period is locked.
It's a waste of time and incoherent to let the user click the button and fill the form and
finally inform him via an access error that the period is locked.

In this commit:
The button was hidden for locked periods.
It Should still be visible to users with the admin access right level.

task-2962093

closes odoo/enterprise#31735

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>