[IMP] payment_sepa_direct_debit: pay once to confirm the mandate
The current way to accept mandates (confirm transactions immediately) is
not considered secure enough since the customer could input any account
number (iban) without confirming that it's their account number.

We introduce a new logic to confirm mandates, an effective payment must
be made before the validation. When a bank statement matching the
transaction data is encoded in the database, the transaction is
confirmed (and the mandate as well, generating the corresponding token).

task-3347338

closes odoo/enterprise#43418

Related: odoo/odoo#126929
Related: odoo/upgrade#4944
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>