[FIX] website_sale_stock_renting: product available to cart
Before this commit, when a rental product is away (rented by a
customer), it is considered as unavailable while it has not been
returned, i.e. by a manual return on SO.
It should only be considered unavailable for the rental period (+ the
security time if one is defined).

Steps to reproduce:
- Select a product that can be rented, of storable type and that is not
configured as "continue selling" when out-of-stock
- Have one unit in stock of this product
- Rental app: rent this product for any period of time > confirm the SO
> validate the pickup
- Go into the website and try to rent for any other period than the one
selected previously
- "Out-of-stock" message

task-3040381
opw-3324198

Part-of: odoo/enterprise#38709