[FIX] project_enterprise: display milestones in the correct period
Currently, milestones planned at the end of a period (week/month/..) are
not displayed correctly in the Gantt view. This commit address this
problem.

Steps
=====
(in a browser not in UTC)
- install module project_enterprise
- Create a task in the Gantt view and a milestone for this task
- Set the deadline of the milestone to the last day of the current month

Issue
=====
The milestone set to the last day of the month is not display in the
good period in the Gantt view. It is displayed in the first day of the
next month.

Cause
=====
There are two issues here:
- The milestone is not display in its actual month: a problem in the
loop exploring each column causes the last slot of the period to not be
checked for milestones.
- The milestone is displayed in the next period: When searching for
milestones for the current period, a datetime, converted to UTC is sent
to the server. This makes no sense as the milestones deadlines are date
and not datetime. When comparing a date to a datetime, the server will
just get rid of the time assosiated to the datetime. But this will cause
problem when the end of the period in the Gantt is converted to
UTC (e.g. July 1st 00:00 can become June 30th 22:00, so a milestone
planned on the 30th of June will in that case appears in the July
period).

Fix
===
- For the first point, an interation is simply added to the loop.
- For the second point, start/end of current Gantt period are first
converted to dates before being used to search milestones.
- On top of that, the detection of a slot at the edge of the gantt
period in the renderer was not working anymore since the conversion in
OWL. This has been fixed as well.

task-3412721

closes odoo/enterprise#48173

X-original-commit: c12ef2e4bef070806a489781a140b2c3882d215b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>