[FIX] web_enterprise: fix rounded custom app icons
=== ISSUE ===

Commit[1] introduced a change on the `border-radius`
property of app icons.

It uses a `rounded` class where it supposed to be `rounded-3`.
This causes inconsistency between these icons and custom app icons
created with Studio.

=== AFTER ===

We reintroduce the changes we introduced in enterprise/#41047.

We add a `rounded-3` class to the `img` tag in order to improve the
consistency.

commit[1]: 4ed891118e576fe623c192a7653cb369ff6936de

task-3326580
part of task-3326263

closes odoo/enterprise#44592

X-original-commit: a2fce27e226d0593c9e3d48a5f95851a00d7c74a
Signed-off-by: Pierre Paridans (app) <app@odoo.com>