[FIX] website_sale_stock_renting: remove picker stock unavailability
The override of `_isInvalidDate` in `website_sale_stock_renting` is
preventing the user to select a day where there is a stock
unavailability. This becomes problematic when the product can be rented
on a hourly basis: a stock unavailability during the day forbids
another renting on that specific day.

Since there is a css class highlighting those potential stock
unavailabilies, it was decided to remove the js restricting the dates
selection of the datetimepicker.

X-original-commit: aadb328ab1a85195d9fe2bb54793ed91279f2b58
Part-of: odoo/enterprise#44158