[FIX] account_reports: fix hierarchy computation
Misc. change: Remove unused array 'account_lines'

Main change:

Currently exactly 1 hierarchy will be calculated for all the input lines.
The account lines are removed from the input lines,
and the lines corresponding to the new hierarchy are then inserted at the end of the lines.
The created hierarchy may span several sections of the report.

This i.e. causes the following problem in the Cash Flow Statement.
    * The accounts relating to opening and closing balance
      respectively are put into the same hierarchy
      (since they represent the same account, just at different times).
    * All the hierarchy lines get added at the very bottom of the report.
      I.e. including: bank and cash accounts relating to opening and
      closing balance and the cash flow entries.
Thus the opening and closing balance lines of the same account
are sibling lines (same parent line) in the report.

The fact that opening and closing balance lines of the same account are sibling lines
in the report leads to a crash in 16.3: Both created lines have the
same ID and the ID is used as key which must be unique.

In this commit the hierarchy computation is changed
such that the sections of the report are respected better.
Each time a new section is encountered the current hierarchy is
rendered and the hierarchy is reset. Thus a new hierarchy will be
created for the following section.

This places the hierarchy lines in the correct spot and prevents the
hierarchy from spanning multiple sections.

This i.e. fixes the problems in the Cash Flow Statement.

closes odoo/enterprise#45338

X-original-commit: 073a83741392cf2ea576b2ef5365844a2c4411c2
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>