[FIX] documents_spreadsheet: fix non-determnistic bundle load
In the document inspector, the spreadsheet bundle is dynamically
loaded.

It appears `loadBundle` takes more than one tick.
Queuing a tick with `nextTick` and awaiting is not enough.

With this commit, we explicitely load and wait the bundle before
it's dynamically loaded in the code under test. Since the loading
is memoized, it's done in a single tick when it's encountered
in the document inspector

runbot error 22383

closes odoo/enterprise#43708

X-original-commit: 87d85fe6ac1d266e195bf0209ff155c4cd5c9958
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>