[FIX] timesheet_grid: prevent starting timer while on input
Steps:
Go to Timesheet > list view.
Type "Mitchell" in the search bar, just as you would do
if you were looking for its timesheets.
Press enter.

Issue:
Well, now the view is filtered, but you just started a timesheet.

Cause:
We listen to key-down events, and if "enter", a timer starts.
Before hand, we have a check, to only do it
when no record is in edition. Yet, it doesn't cover this case.

Fix:
Add the condition that the event doesn't target an input or a text area.

task-3413333

closes odoo/enterprise#43936

X-original-commit: 7051e0d21d6b40aa50097d2cb7be56c001f1c1b9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>