[FIX] helpdesk: make tests (demo) data independent
Before this commit, the hr_holiday tests used base_company to setup the
tests. This resulted in the tests interacting with the demo data that
were linked to this company, eg stress days.

With this commit, the tests are now independent of this data, and will
therefore not be affected by any changes to it.

The company in the helpdesk test suite is set the company of
self.env.user, so that it works together nicely with hr_holidays in
the helpdesk_holidays tests.

closes odoo/enterprise#43717

Related: odoo/odoo#127415
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>