[IMP] appointment: add resource on events
Currently appointment events can use multiple
resources. However this makes handling them
more complicated in cases where we do not care
about handling multiple resources.

We add an m2o field on calendar events for booking lines
as well as a few inverse and compute methods to allow
users to manipulate the booking line values when
there is only one. We also add some infered default values
when creating the record with a default resource to
make creation as easy as possible.

As part of these usability compute methods we make
the used capacity a computed field such that users
can just enter the reserved capacity when relevant.

task-3101589

Part-of: odoo/enterprise#37312