[FIX] delivery_sendcloud: return label attachment as a list
When a user attempts to get the return label using the
`sendcloud_get_return_label` method,  they will face the ValueError:
Posting a message should receive attachments records as a list of IDs.

The `_prepare_track_message_docs` method returns the doc_ids as a recordset.
However, the `message_post_with_source` method expects the parameter
attachment_ids to be provided as a list of IDs rather than a recordset.

To resolve this issue, the parameter should be passed as a list of
attachment IDs instead.

Sentry-4295276140

closes odoo/enterprise#44186

X-original-commit: cc0d9b51530b8ca071de3e7d0e25e56bc5111204
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Parth Solanki (paso) <paso@odoo.com>