[FIX] mrp_workorder: tablet form background
Prior to this commit, the background of `o_workorder_tablet_form` was
white due to the inheritance of `o_form_nosheet`.
The content, which was also white, was not visible.

This commit fixes this issue by removing the white background for the
`o_workorder_tablet_form`.

task-3414400
Part of task-3326263

closes odoo/enterprise#43838

X-original-commit: 847f6a23a32821fd7fb5e4c0c1e2ab4a3d150cf1
Signed-off-by: Pierre Paridans (app) <app@odoo.com>