[IMP] project_enterprise: add empty line for assignees in gantt view
In the project.task gantt view, an empty line is displayed for
assignees to a planned and opened task, but only for the next period. Ideally,
if a planned and opened task exists in the project, it would be better to have
an empty line displayed for its assignees in the whole gantt view, not only for
the next period.

This commit will change the group expand's domain to include tasks that
have a planned_date_begin and planned_date_end set, thus including
assignees who have planned tasks.

task-3414550

Part-of: odoo/enterprise#43757