[FIX] account_bank_statement_import_csv: import balance correctly.
When importing bank statement lines from a CSV file, using debit, credit and
balance fields (with at least one of the column for debit or credit before
the column for balance), the 'balance_start' and 'balance_end_real' computed
by the _parse_import_data function will be incorrect.

This is because of the order of the operations in that function.
First, the imported data ('imported_data') is parsed into a variable
named 'data'.
Then, the 'imported_data' will be checked for 'debit' and 'credit' column. If
they are found, they are processed then the columns are deleted from
'imported_data'.
Lastly, 'imported_data' is checked for the 'balance' column. If it is found,
its index in 'imported_data' is stored and then accessed in 'data'. This is the
incorrect part. Since we deleted two columns from 'imported_data', the index
for balance is not the same as in 'data' anymore (since it still
contains 'debit' and/or 'credit' before 'balance').

As discussed with POWA, the correct fix would be to delete the debit/credit
column after processing the balance column in the imported data, so the indexes
remain the same.

opw-3380610

closes odoo/enterprise#43766

Signed-off-by: Quentin De Paoli <qdp@odoo.com>