[FIX] web_enterprise, web_studio: fix UI issues
This commit fixes various issues with the app switcher when using Studio.

Here is a list of the different issues handled in this commit :

===============================
Edit icon button is not aligned
===============================

When trying to edit an app icon, the pencil icon is not correctly placed.
It should be displayed in the top right corner of the tile and not in the
center.

============================
Custom icon preview modal UI
============================

If you open Studio and click on the pencil icon to edit an icon, there are
some issues with the modal displayed.

> The preview image is not centered inside the modal.

> The `Confirm` primary button use the same color for the `background`
and the text in active state, causing readability issues.

======================
Studio navbar contrast
======================

Still on the App Switcher while Studio is activated, the `Customizations`
button at the top left of the screen in black on a dark background, which
is not readable at all. There is also an issue with the pencil icon being
placed weirdly to the icon container, it should be aligned to the top right
corner.

=================================
Custom uploaded icon aspect-ratio
=================================

When an icon is not squared, its container would fit the image size while
it should normally force the icon to be squared. This would lead to a
rectangle icon in the middle of squared icon.

========================================================
Custom uploaded icon aspect-ratio inside a module navbar
========================================================

While fixing the issues above, I found out that the `aspect-ratio` also
affects the icon rendered inside a module navbar. If you upload an icon
with rectangle dimensions and open a module, the icon in the navbar would
often overflows the name of the module and hide it. This is also not the
expected render since we made the icon squared on the app switcher.

=> To handle this, we constrain the width of the image to the navbar
height in order to obtain a squared image. We first tried to fix this
using an aspect-ratio but it was affecting the regular icons which
sometimes made their edges cut.

========================================================
Field properties menu
========================================================

Since we apply changes inside Studio, we also apply the same changes we
did on the Menu Editor in Website, by changing the background of the
buttons. There was also an issue with the label of the icons which were
the same even if the buttons had different purposes.

task-3418936
part of task-3326263

Forward-port-of: #43857
X-original-commit: 6a21d972d2eb822e5bb504ec1a734dacfc5ef085
Part-of: odoo/enterprise#48174